Retro-fitting tests to the Results
I'm not sure here about how to best expose the subject.

We could either:

1) pass out the subject as a second parameter in the passed / failed method
2) have another #describe_subject_to method
3) just leave it as an attribute that you can get if you want it

As it is, I've gone for having the describe_to on the result delegate to
the describe_to on the subject, so if you ask a result to describe
itself, it will tell you about itself, then about its subject.

@tooky WDYT?