Modify test to be more integratey
@tom025 I'm wondering whether this test doesn't belong on the runner
now, as the runner is still calling `#skip`, and we use polymorphism
lower down to swap in a mapping that can't be skipped.

I'm not sure where the best place is for this test. HookCompiler has
it's own specs, but this is really about the behaviour of HookCompiler
and Runner acting in concert. WDYT?