Be explicit about which visitor will walk the next part of the tree.
RSpec 3.0 is unable to correctly report blocks with optional paramters
that are yielded by stubbed methods.

This was causing a number of failures in this test suite.

We had the option to continue walking the tree with the original
visitor, or to pass in a new visitor to the `descend` block.

I've removed that option - if the original visitor wants to continue
walking the tree, and not use a child visitor, it can just pass itself
in.

I think this makes the describe_to code a little easier to parse?

cc: @tom025 @mattwynne