Use the location of the row for instantiated outline steps
When steps are instantiated from outline steps, they get the location
of the row (instead of the location of the outline step). To make sure
that a test case of an example row still matches the steps of the
Scenario Outline, the instantiated steps know of the outline step that
created them.