Simplify API for filter chain in core
Added a new method to the filter protocol - every filter now has to
implement #with_receiver which returns an instance of the filter
with the given receiver.

I've changed each of our filters' constructors to use nil as the
receiver by default. I considered writing a special null object, but
I decided it's not neccessary. When the receiver is nil, the filter is
in an 'unready' state and it would be fine for it to blow up with
NoMethodErrors if you tried to use it.

cc @tooky - WDYT?