Remove the Mapper DSL
I've removed most of the tests that refer to hooks, except the around
hooks in the core_spec.rb, not sure this is tested anywhere else... but
this probably isn't the right place. I've left an AroundHook
implementation in core because of that test, but we have one in the
front end too?

Also updated README to construct a filter to activate steps, not sure if it
tells the story?